+++ pkg/scalers/authentication/authentication_types.go
@@ -16,7 +16,7 @@ const (
 	BearerAuthType Type = "bearer"
 	// CustomAuthType is an auth type using a custom header
 	CustomAuthType Type = "custom"
-	// OAuthType is a auth type using a oAuth2
+	// OAuthType is an auth type using a oAuth2
 	OAuthType Type = "oauth"
 )